# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is list API's to be used in Quiz App.",
        "title": "Quiz App API's",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/users/check-user/{phone}": {
            "get": {
                "description": "generateing OTP for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User AUth"
                ],
                "summary": "generateing OTP for the user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user phone number",
                        "name": "phone",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.OTPResponse"
                        }
                    }
                }
            }
        },
        "/users/edit-user-profile": {
            "post": {
                "description": "This API edit user profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User AUth"
                ],
                "summary": "This API edit user profile",
                "parameters": [
                    {
                        "description": "this is user info json",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.profileInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.editProfileResponse"
                        }
                    }
                }
            }
        },
        "/users/get-user-wallet-details/{user_id}": {
            "get": {
                "description": "Fetches user wallet and transaction details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Wallet"
                ],
                "summary": "Get user wallet details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIResponse"
                        }
                    }
                }
            }
        },
        "/users/get-user/{id}": {
            "get": {
                "description": "This API will provide user info bu id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User AUth"
                ],
                "summary": "This API will provide user info bu id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "this is user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.userInfo"
                        }
                    }
                }
            }
        },
        "/users/initiate-user-transaction": {
            "post": {
                "description": "This API will make user transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Wallet"
                ],
                "summary": "This API will make user transactions",
                "parameters": [
                    {
                        "description": "this is transaction info json",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.requestTransaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.transactionResponse"
                        }
                    }
                }
            }
        },
        "/users/verify-otp/{userId}/{otp}": {
            "get": {
                "description": "This API will verify user OTP with userId",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User AUth"
                ],
                "summary": "This API will verify user OTP with userId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user app Id",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "user otp",
                        "name": "otp",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.verifyOtpResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.APIResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "$ref": "#/definitions/controllers.WalletDetails"
                },
                "message": {
                    "type": "string",
                    "example": "User wallet details found."
                },
                "status": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
...
        "controllers.OTPResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "object",
                    "properties": {
                        "otp": {
                            "type": "string",
                            "example": "8162"
                        },
                        "userId": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "New user created"
                },
                "status": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
        "controllers.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 20
                },
                "created": {
                    "type": "string",
                    "example": "2024-12-18T23:04:50+05:30"
                },
                "title": {
                    "type": "string",
                    "example": "Deposit"
                },
                "transaction_type": {
                    "type": "string",
                    "example": "CREDIT"
                }
            }
        },
        "controllers.WalletDetails": {
            "type": "object",
            "properties": {
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.Transaction"
                    }
                },
                "wallet": {
                    "type": "number"
                }
            }
        },
...
        "controllers.editProfileResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "User info updated successfully."
                },
                "status": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
...
        "controllers.profileInfo": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "sn@gmail.com"
                },
                "gender": {
                    "type": "string",
                    "example": "Male"
                },
                "image": {
                    "type": "string",
                    "example": "url-of-the-image"
                },
                "name": {
                    "type": "string",
                    "example": "Shivam Nagpal"
                },
                "phone": {
                    "type": "string",
                    "example": "0987656"
                },
                "userId": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
...
        "controllers.requestTransaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 200
                },
                "transaction_type": {
                    "type": "string",
                    "example": "DEBIT"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "controllers.transactionResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "User info updated successfully."
                },
                "status": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
        "controllers.userInfo": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "object",
                    "properties": {
                        "ID": {
                            "type": "integer",
                            "example": 3
                        },
                        "created": {
                            "type": "string",
                            "example": "2024-12-10T07:04:37Z"
                        },
                        "email": {
                            "type": "string",
                            "example": "shivam@gmail.com"
                        },
                        "name": {
                            "type": "string",
                            "example": "Shivam"
                        },
                        "otp": {
                            "type": "string",
                            "example": "8162"
                        },
                        "phone": {
                            "type": "string",
                            "example": "9144"
                        },
                        "register": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "User logged in successfully"
                },
                "status": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
        "controllers.verifyOtpResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "object",
                    "properties": {
                        "ID": {
                            "type": "integer",
                            "example": 3
                        },
                        "created": {
                            "type": "string",
                            "example": "2024-12-10T07:04:37Z"
                        },
                        "email": {
                            "type": "string",
                            "example": "shivam@gmail.com"
                        },
                        "name": {
                            "type": "string",
                            "example": "Shivam"
                        },
                        "otp": {
                            "type": "string",
                            "example": "8162"
                        },
                        "phone": {
                            "type": "string",
                            "example": "9144"
                        },
                        "register": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "User logged in successfully"
                },
                "status": {
                    "type": "string",
                    "example": "1"
                }
            }
        }
    }
}